% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stuff_to_json.R
\name{awards_to_json}
\alias{awards_to_json}
\title{Build Awards JSON}
\usage{
awards_to_json(matchup_data, player_data, max_week, team_photos)
}
\arguments{
\item{matchup_data}{team-level matchup data from \code{get_team_matchups()}}

\item{player_data}{player-level data from \code{get_all_matchups_data()}}

\item{max_week}{the current week number}

\item{team_photos}{team photo data from \code{get_team_photos()}}
}
\value{
a JSON string
}
\description{
Computes best-ball lineups, creates the awards table, and returns the result
as a JSON string.
}
