% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motw.R
\name{create_danger_chart}
\alias{create_danger_chart}
\title{Plot the MotW Danger Chart}
\usage{
create_danger_chart()
}
\value{
a ggplot2 object
}
\description{
Reads the main MotW simulation CSV (\code{currentMotW/matchup_tree.csv}) and
returns a stacked bar chart showing each team's probability of having to take
shots by future week.
}
\details{
Requires \code{currentMotW/matchup_tree.csv} to exist (generated by
  \code{motw.py}).
}
