% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{create_PF_PA_scatter}
\alias{create_PF_PA_scatter}
\title{Plot Points For vs Points Against Scatter}
\usage{
create_PF_PA_scatter(leaderboard, team_photos)
}
\arguments{
\item{leaderboard}{leaderboard data frame from \code{create_leaderboard()}}

\item{team_photos}{team photo data from \code{get_team_photos()}}
}
\value{
a ggplot2 object
}
\description{
Scatter plot showing each team's total PF vs PA with team logos (or name
labels as fallback), and quadrant annotations for luck/skill.
}
