% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{create_rank_chart}
\alias{create_rank_chart}
\title{Plot Weekly Team Rankings}
\usage{
create_rank_chart(matchup_data, team_photos)
}
\arguments{
\item{matchup_data}{team-level matchup data from \code{get_team_matchups()}}

\item{team_photos}{team photo data from \code{get_team_photos()}}
}
\value{
a ggplot2 object
}
\description{
Bump chart showing each team's weekly rank (by points scored), with team
logos or name labels at each point and W/L indicators.
}
