% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{create_shots_dist}
\alias{create_shots_dist}
\title{Plot the MotW Shots Distribution}
\usage{
create_shots_dist(motw_data, max_week, color_1, color_2)
}
\arguments{
\item{motw_data}{MotW-augmented matchup data from \code{add_motw_to_matchups()}}

\item{max_week}{the current week number}

\item{color_1}{fill colour for MotW losers}

\item{color_2}{fill colour for non-MotW losers}
}
\value{
a ggplot2 object
}
\description{
Stacked bar chart of how many potential shots losing teams would take each
week, split by whether they were in a MotW game.
}
