% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stuff_to_json.R
\name{efficiency_to_json}
\alias{efficiency_to_json}
\title{Build Efficiency JSON}
\usage{
efficiency_to_json(matchup_data, player_data, max_week)
}
\arguments{
\item{matchup_data}{team-level matchup data from \code{get_team_matchups()}}

\item{player_data}{player-level data from \code{get_all_matchups_data()}}

\item{max_week}{the current week number}
}
\value{
a JSON string
}
\description{
Computes best-ball vs actual points for the current week and returns as JSON.
}
