% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awards_table.R
\name{find_top_player}
\alias{find_top_player}
\title{Title Find the Top Player at a Position in a Week}
\usage{
find_top_player(pos, data, current_week)
}
\arguments{
\item{pos}{Position to look for}

\item{data}{player data}

\item{current_week}{Week to look at}
}
\description{
Title Find the Top Player at a Position in a Week
}
\examples{
find_top_player("QB", player_data, 2)
}
