% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper-etl.R
\name{get_all_matchups_data}
\alias{get_all_matchups_data}
\title{Title Get All Matchup Data up to the Current Week}
\usage{
get_all_matchups_data(max_week, league_id, file_path)
}
\arguments{
\item{max_week}{the most recent week of the NFL season}

\item{league_id}{Sleeper league id to get data from}

\item{file_path}{a file path to read the csv file of player information}
}
\description{
Title Get All Matchup Data up to the Current Week
}
\examples{
get_all_matchups_data(2, 996445270105714688, "sleeper_players.csv")
}
