% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper-etl.R
\name{get_all_starters_data}
\alias{get_all_starters_data}
\title{Title Get Weekly Starters}
\usage{
get_all_starters_data(max_week, league_id)
}
\arguments{
\item{max_week}{the most recent week of the NFL season}

\item{league_id}{Sleeper league id to get data from}
}
\description{
Title Get Weekly Starters
}
\examples{
get_all_starters_data(2, 996445270105714688)
}
