% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper-etl.R
\name{get_players_data}
\alias{get_players_data}
\title{Title Get Player Data From Sleeper}
\usage{
get_players_data(update_players = FALSE, file_path)
}
\arguments{
\item{update_players}{whether or not to update players}

\item{file_path}{a file path to read/write the csv file of player information}
}
\description{
Title Get Player Data From Sleeper
}
\examples{
get_players_data("sleeper_players.csv")
get_players_data(update_players = TRUE, "sleeper_players.csv")
}
