% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper-etl.R
\name{get_team_matchups}
\alias{get_team_matchups}
\title{Title Get Matchups for Each Team Each Week}
\usage{
get_team_matchups(player_data)
}
\arguments{
\item{player_data}{result of get_all_matchups_data}
}
\description{
Title Get Matchups for Each Team Each Week
}
\examples{
get_team_matchups(get_all_matchups_data(2, 996445270105714688, "sleeper_players.csv"))
}
