% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper-etl.R
\name{get_transactions_sleeper}
\alias{get_transactions_sleeper}
\title{Get Transactions for a Sleeper League Round}
\usage{
get_transactions_sleeper(league_id, round)
}
\arguments{
\item{league_id}{Sleeper league id to get data from}

\item{round}{the week/round number to retrieve transactions for (numeric)}
}
\value{
a data frame of transactions for the given round, or nothing if no
  data is found
}
\description{
Queries the Sleeper API for all transactions (adds, drops, trades) in a
given week/round of a league.
}
