% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper-etl.R
\name{get_week_matchup_data}
\alias{get_week_matchup_data}
\title{Title Get Matchup Data for a Given Week}
\usage{
get_week_matchup_data(week_number, league_id)
}
\arguments{
\item{week_number}{NFL week to get matchup data for}

\item{league_id}{Sleeper league id to get data from}
}
\description{
Title Get Matchup Data for a Given Week
}
\examples{
get_week_matchup_data(1, 996445270105714688)
}
