% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stuff_to_json.R
\name{matchup_info_to_json}
\alias{matchup_info_to_json}
\title{Build Matchup Info JSON}
\usage{
matchup_info_to_json(matchup_data, team_photos)
}
\arguments{
\item{matchup_data}{team-level matchup data from \code{get_team_matchups()}}

\item{team_photos}{team photo data from \code{get_team_photos()}}
}
\value{
a JSON string
}
\description{
Adds per-week statistics (average, median, max, min, margin of victory) and
color scales to the matchup data, then returns as JSON.
}
