% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stuff_to_json.R
\name{power_rankings_to_json}
\alias{power_rankings_to_json}
\title{Build Power Rankings JSON}
\usage{
power_rankings_to_json(matchup_data, max_week, number_of_teams)
}
\arguments{
\item{matchup_data}{team-level matchup data from \code{get_team_matchups()}}

\item{max_week}{the current week number}

\item{number_of_teams}{total number of teams in the league}
}
\value{
a JSON string
}
\description{
Creates power rankings with team ability, strength of schedule, and color
columns, returning as JSON.
}
