% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stuff_to_json.R
\name{schedule_comparison_to_json}
\alias{schedule_comparison_to_json}
\title{Build Schedule Comparison JSON}
\usage{
schedule_comparison_to_json(matchup_data, team_photos)
}
\arguments{
\item{matchup_data}{team-level matchup data from \code{get_team_matchups()}}

\item{team_photos}{team photo data from \code{get_team_photos()}}
}
\value{
a JSON string
}
\description{
For every pair of teams, calculates the head-to-head record as if they had
played each other every week (best record, worst record, actual record) and
returns as a nested JSON string.
}
