% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recap.R
\name{summarize_motw_given}
\alias{summarize_motw_given}
\title{Summarize shots/dogs given out by each MotW winner}
\usage{
summarize_motw_given(motw_data)
}
\arguments{
\item{motw_data}{result of \code{add_motw_to_matchups}}
}
\value{
data frame with one row per team: \code{team_name},
  \code{shots_dogs_given_out}
}
\description{
Summarize shots/dogs given out by each MotW winner
}
