% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recap.R
\name{summarize_pt_diff}
\alias{summarize_pt_diff}
\title{Summarize average point differential, close games, and blowouts per team}
\usage{
summarize_pt_diff(all_matchups)
}
\arguments{
\item{all_matchups}{matchup data}
}
\value{
data frame with one row per \code{manager_id}: \code{pt_diff_1},
  \code{pt_diff_0}, \code{close_games_1}, \code{close_games_0},
  \code{blowouts_1}, \code{blowouts_0}, and associated color columns
}
\description{
Summarize average point differential, close games, and blowouts per team
}
