% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recap.R
\name{summarize_weekly_rank}
\alias{summarize_weekly_rank}
\title{Count weeks each team was the best/worst scorer in the league}
\usage{
summarize_weekly_rank(all_matchups)
}
\arguments{
\item{all_matchups}{matchup data}
}
\value{
data frame with one row per \code{manager_id}: best/worst team week
  counts and associated color columns
}
\description{
Count weeks each team was the best/worst scorer in the league
}
