% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recap.R
\name{summarize_win_loss_extremes}
\alias{summarize_win_loss_extremes}
\title{Summarize best and worst win/loss scores per team}
\usage{
summarize_win_loss_extremes(all_matchups)
}
\arguments{
\item{all_matchups}{matchup data}
}
\value{
data frame with one row per team: \code{manager_id}, \code{team_name},
  score extremes, and associated color columns
}
\description{
Summarize best and worst win/loss scores per team
}
