% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awards_table.R
\name{top_player_award}
\alias{top_player_award}
\title{Title Create the Top Player Award}
\usage{
top_player_award(pos, data, current_week, award_name)
}
\arguments{
\item{pos}{position for award}

\item{data}{player data}

\item{current_week}{week of award}

\item{award_name}{what to name the award}
}
\description{
Title Create the Top Player Award
}
\examples{
top_player_award("QB", player_data, 2, "Anti-Russ")
}
