% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleeper-etl.R
\name{update_player_data}
\alias{update_player_data}
\title{Update Player Data}
\usage{
update_player_data(file_path)
}
\arguments{
\item{file_path}{a file path to save the csv file of player information to}
}
\description{
Update Player Data
}
\examples{
update_player_data("sleeper_players.csv")
}
