% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stuff_to_json.R
\name{write_json_to_file}
\alias{write_json_to_file}
\title{Write JSON Data to a File}
\usage{
write_json_to_file(json_data, file_path)
}
\arguments{
\item{json_data}{a JSON string (e.g. from \code{jsonlite::toJSON()})}

\item{file_path}{destination file path}
}
\value{
invisibly returns \code{NULL}; called for its side effect
}
\description{
Writes a JSON string to the specified file path and prints a confirmation
message.
}
