% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motw.R
\name{add_motw_to_matchups}
\alias{add_motw_to_matchups}
\title{Title Matchup of the Week}
\usage{
add_motw_to_matchups(matchup_data, week_1_matchup_id, max_week)
}
\arguments{
\item{matchup_data}{matchup data from sleeper}

\item{week_1_matchup_id}{matchup id for the MotW for week 1}

\item{max_week}{the final week to calculate MotW for}
}
\description{
Title Matchup of the Week
}
\examples{
add_motw_to_matchups(matchup_data = all_matchups, week_1_matchup_id = 1, max_week = current_week)
}
